Base implementation of clean architecture. Separation of concerns into data, domain and application layers
