--- app/src/main/java/io/chicago/art/artwork/dto/api/ArtWorks.kt
@@ -1,3 +1,0 @@
-package io.chicago.art.artwork.dto.api
-
-data class ArtWorks(val pagination: Pagination, val data: List<Data>, val config: Config)
