--- app/src/main/java/io/chicago/art/artwork/dto/api/Config.kt
@@ -1,3 +1,0 @@
-package io.chicago.art.artwork.dto.api
-
-data class Config(val iiif_url: String)
