--- app/src/main/java/io/chicago/art/artwork/dto/api/Data.kt
@@ -1,3 +1,0 @@
-package io.chicago.art.artwork.dto.api
-
-data class Data(val title: String, val image_id: String)
