--- app/src/main/java/io/chicago/art/artwork/dto/api/Pagination.kt
@@ -1,8 +1,0 @@
-package io.chicago.art.artwork.dto.api
-
-data class Pagination(
-    val total: Int,
-    val limit: Int,
-    val offset: Int,
-    val total_pages: Int,
-    val current_page: Int)
