--- app/src/main/java/io/chicago/art/viewstate/DefaultViewState.kt
@@ -1,10 +1,0 @@
-package io.chicago.art.viewstate
-
-sealed class DefaultViewState<T> {
-
-    class Loading<T> : DefaultViewState<T>()
-
-    class Error<T> : DefaultViewState<T>()
-
-    class Success<T>(val result: T) : DefaultViewState<T>()
-}
