--- app/src/main/java/io/chicago/art/webservice/WebService.kt
@@ -1,13 +1,0 @@
-package io.chicago.art.webservice
-
-import io.chicago.art.artwork.dto.api.ArtWorks
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-const val baseUrl = "https://api.artic.edu/api/v1/"
-
-interface WebService {
-
-    @GET("artworks")
-    suspend fun getArtworks(@Query("limit") limit: Int = 25): ArtWorks
-}
