+++ app/src/main/res/layout/artwork_fragment.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="artwork"
-            type="io.chicago.art.artwork.dto.ui.ArtWorkUIItem" />
+            type="io.chicago.data.entities.ArtWorkUIItem" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
