+++ app/src/main/res/navigation/nav_graph.xml
@@ -4,7 +4,7 @@
     app:startDestination="@id/collectionContainerFragment">
     <fragment
         android:id="@+id/collectionContainerFragment"
-        android:name="io.chicago.art.artwork.view.ArtWorkContainerFragment"
+        android:name="io.chicago.art.view.ArtWorkContainerFragment"
         android:label="CollectionContainerFragment" >
     </fragment>
 </navigation>
