+++ domain/src/main/java/io/chicago/domain/entities/ArtWorks.kt
@@ -1,0 +1,3 @@
+package io.chicago.domain.entities
+
+data class ArtWorks(val pagination: Pagination, val data: List<Data>, val config: Config)
