+++ domain/src/main/java/io/chicago/domain/entities/Config.kt
@@ -1,0 +1,3 @@
+package io.chicago.domain.entities
+
+data class Config(val iiif_url: String)
