+++ domain/src/main/java/io/chicago/domain/entities/Data.kt
@@ -1,0 +1,3 @@
+package io.chicago.domain.entities
+
+data class Data(val title: String, val image_id: String)
