+++ domain/src/main/java/io/chicago/domain/usecases/GetArtWorkUseCase.kt
@@ -1,0 +1,7 @@
+package io.chicago.domain.usecases
+
+import io.chicago.domain.repositories.ArtWorkRepository
+
+class GetArtWorkUseCase(private val artWorkRepository: ArtWorkRepository) {
+    suspend operator fun invoke() = artWorkRepository.fetchArtworks()
+}
