+++ settings.gradle
@@ -1,2 +1,4 @@
 rootProject.name = "ChicagoArt"
 include ':app'
+include ':domain'
+include ':data'
